Update integer parsing and field creation to use int64
This commit modifies the parseIntField function to read 64-bit integers from the input, ensuring compatibility with larger integer values. Additionally, the CreateFieldFromConstant function is updated to use int64 for parsing constants, aligning with recent changes in integer handling across the codebase. These updates enhance the consistency and range of integer support in the system.

+++ pkg/types/parser.go
@@ -57,7 +57,7 @@ func parseIntField(r io.Reader, maxSize uint32) (*IntField, error) {
 		return nil, err
 	}
 
-	value := int64(binary.BigEndian.Uint32(bytes))
+	value := int64(binary.BigEndian.Uint64(bytes))
 	return NewIntField(value), nil
 }
